Cleanup

+++ scripts/generate_ocr_files.py
@@ -19,94 +19,92 @@ def generate_ocr_json(path, ocr_path):
         if not ocr_path.exists():
             os.mkdir(ocr_path)
             print('_ocr folder created')
-            with open(path, 'r', encoding='utf-8') as f:
-                html = f.read()
-                soup = BeautifulSoup(html, 'html.parser')
+        with open(path, 'r', encoding='utf-8') as f:
+            html = f.read()
+            soup = BeautifulSoup(html, 'html.parser')
+            pages = soup.find_all(class_="page")
+            for page in pages:
+                page_container = page.find(class_="pageContainer")
+                styles = page_container["style"].split(";")[:-1]
+                for style in styles:
+                    key, value = style.strip().split(":")
+                    if (key == "width"):
+                        img_width = int(value)
+                    if (key == "height"):
+                        img_height = int(value)
+                    if (key == "background-image"):
+                        background_image = value
 
-                pages = soup.find_all(class_="page")
-                for page in pages:
-                    page_container = page.find(class_="pageContainer")
-                    styles = page_container["style"].split(";")[:-1]
-                    for style in styles:
-                        key, value = style.strip().split(":")
-                        if (key == "width"):
-                            img_width = int(value)
-                        if (key == "height"):
-                            img_height = int(value)
-                        if (key == "background-image"):
-                            background_image = value
+                image_name = unquote(re.findall(
+                    '"([^"]*)"', background_image)[0])
+                image_path = Path(image_name)
+                folder_name = image_path.parent
+                folder_path = Path(os.path.join(ocr_path, folder_name))
 
-                    image_name = unquote(re.findall(
-                        '"([^"]*)"', background_image)[0])
-                    image_path = Path(image_name)
-                    folder_name = image_path.parent
-                    blocks = []
-                    text_boxes = page.find_all(class_="textBox")
-                    for text_box in text_boxes:
-                        if text_box is not None:
-                            text_box_styles = text_box["style"].split(";")[:-1]
-                            for text_box_style in text_box_styles:
-                                text_box_style_key, text_box_style_value = text_box_style.strip().split(":")
-                                writing_mode = None
-                                if (text_box_style_key == "width"):
-                                    text_box_width = int(text_box_style_value)
-                                if (text_box_style_key == "height"):
-                                    text_box_height = int(text_box_style_value)
-                                if (text_box_style_key == "left"):
-                                    xmin = int(text_box_style_value)
-                                if (text_box_style_key == "top"):
-                                    ymin = int(text_box_style_value)
-                                if (text_box_style_key == 'writing-mode'):
-                                    writing_mode = text_box_style_value
-                                if (text_box_style_key == 'font-size'):
-                                    font_size = text_box_style_value
-                            xmax = xmin + text_box_width
-                            ymax = ymin + text_box_height
-                            box = [xmin, ymin, xmax, ymax]
-                            if writing_mode is not None:
-                                vertical = True
-                            else:
-                                vertical = False
+                if not folder_path.exists():
+                    os.makedirs(folder_path)
 
-                            lines = []
-                            p_tags = text_box.find_all('p')
-                            for p in p_tags:
-                                lines.append(p.text.strip())
-                        block = {
-                            "box": box,
-                            "vertical": vertical,
-                            "font_size": float(font_size.replace("px", "")),
-                            "lines": lines
-                        }
-                        blocks.append(block)
+                blocks = []
+                text_boxes = page.find_all(class_="textBox")
+                for text_box in text_boxes:
+                    if text_box is not None:
+                        text_box_styles = text_box["style"].split(";")[:-1]
+                        for text_box_style in text_box_styles:
+                            text_box_style_key, text_box_style_value = text_box_style.strip().split(":")
+                            writing_mode = None
+                            if (text_box_style_key == "width"):
+                                text_box_width = int(text_box_style_value)
+                            if (text_box_style_key == "height"):
+                                text_box_height = int(text_box_style_value)
+                            if (text_box_style_key == "left"):
+                                xmin = int(text_box_style_value)
+                            if (text_box_style_key == "top"):
+                                ymin = int(text_box_style_value)
+                            if (text_box_style_key == 'writing-mode'):
+                                writing_mode = text_box_style_value
+                            if (text_box_style_key == 'font-size'):
+                                font_size = text_box_style_value
+                        xmax = xmin + text_box_width
+                        ymax = ymin + text_box_height
+                        box = [xmin, ymin, xmax, ymax]
+                        if writing_mode is not None:
+                            vertical = True
+                        else:
+                            vertical = False
 
-                    page = {
-                        "version": "0.1.6",
-                        "img_width": img_width,
-                        "img_height": img_height,
-                        "blocks": blocks,
+                        lines = []
+                        p_tags = text_box.find_all('p')
+                        for p in p_tags:
+                            lines.append(p.text.strip())
+                    block = {
+                        "box": box,
+                        "vertical": vertical,
+                        "font_size": float(font_size.replace("px", "")),
+                        "lines": lines
                     }
+                    blocks.append(block)
 
-                    folder_path = Path(os.path.join(ocr_path, folder_name))
+                page = {
+                    "version": "0.1.6",
+                    "img_width": img_width,
+                    "img_height": img_height,
+                    "blocks": blocks,
+                }
 
-                    if not folder_path.exists():
-                        os.makedirs(folder_path)
+                json_file_name = os.path.join(
+                    folder_path, f'{image_path.stem}.json')
 
-                    json_file_name = os.path.join(
-                        folder_path, f'{image_path.stem}.json')
+                json_dump = json.dumps(
+                    page, ensure_ascii=False).encode('utf8').decode()
 
-                    json_dump = json.dumps(
-                        page, ensure_ascii=False).encode('utf8').decode()
-
-                    with open(json_file_name, 'w', encoding='utf-8') as json_file:
-                        json_file.write(json_dump)
-                        json_file.close()
-                return folder_name
-        else:
-            print('_ocr folder already exists')
+                with open(json_file_name, 'w', encoding='utf-8') as json_file:
+                    json_file.write(json_dump)
+                    json_file.close()
+            return os.path.dirname(folder_name)
 
 
 def main():
+    print('----------------------------------------------')
     if not path.exists():
         print('Path does not exist')
         exit()
@@ -122,12 +120,19 @@ def main():
         print(f'Generating ocr json for all html files in {path.name}')
         ocr_path = Path(os.path.join(path, ocr_dir))
         for p in Path(path).expanduser().absolute().iterdir():
-            generate_ocr_json(p, ocr_path)
-            if remove_originals and run_mokuro and p.is_file() and p.suffix == '.html':
-                os.remove(p)
+            if p.suffix == '.mokuro':
+                print(f'{path.name} is already processed')
+                return
+
+        if not ocr_path.exists():
+            for p in Path(path).expanduser().absolute().iterdir():
+                generate_ocr_json(p, ocr_path)
+                if remove_originals and run_mokuro and p.is_file() and p.suffix == '.html':
+                    os.remove(p)
         if run_mokuro:
             print(f'Reprocessing directory {path.name}')
             run(parent_dir=path, disable_confirmation=True, legacy_html=False)
+
     else:
         print(f'Generating _ocr files')
         folder_name = generate_ocr_json(path, ocr_dir)
@@ -160,7 +165,7 @@ if __name__ == '__main__':
 
     if parent_dir:
         print(f'Processing every directory within {path.name}')
-        for p in path.expanduser().absolute().iterdir():
+        for p in sorted(path.expanduser().absolute().iterdir()):
             if p.is_dir():
                 path = p
                 main()
